Added filter search
Hex-Utils: can now handle filters, save board as image. move history allows to mark the order in which the pieces were placed.
seize_isterminal: determines if the set of pieces on the board are connected.
seize: generate a pickle file which contains all the possible filters for the specified board.